refactor: renamed constants.ts -> ALObject\RegexPatterns.ts
