+++ extension/src/ALObject/RegexPatterns.ts
